requirements.txt
absl-py==0.9.0
astunparse==1.6.3
autokeras==1.0.5
cachetools==4.0.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
gast==0.3.3
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageio==2.9.0
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.3.1
networkx==2.4
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.4.0.40
opt-einsum==3.2.0
packaging==20.4
pandas==1.1.0
Pillow==7.2.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.20.0
scipy==1.4.1
six==1.14.0
sklearn==0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
# tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.8.13
tqdm==4.48.2
urllib3==1.25.8
Werkzeug==1.0.0
wincertstore==0.2
wrapt==1.12.1
